+++ src/renderer/src/components/right-sidebar/source-control/sync/use-branch-compare-refresh-triggers.ts
@@ -1,0 +1,88 @@
+import { useEffect, useRef, type MutableRefObject } from 'react'
+import type { GitUpstreamStatus } from '../../../../../../shared/git-status-types'
+import {
+  shouldRefreshBranchCompareForRemoteStatus,
+  shouldRefreshBranchCompareForStatusHead,
+  type BranchCompareRemoteStatusSnapshot,
+  type BranchCompareStatusHeadSnapshot
+} from './compare-summary'
+
+export function useBranchCompareRefreshTriggers({
+  activeWorktreeId,
+  worktreePath,
+  compareBaseRef,
+  isFolder,
+  isBranchVisible,
+  activeGitStatusHead,
+  remoteStatus,
+  refreshBranchCompareRef
+}: {
+  activeWorktreeId: string | null
+  worktreePath: string | null
+  compareBaseRef: string | null
+  isFolder: boolean
+  isBranchVisible: boolean
+  activeGitStatusHead: string | null
+  remoteStatus: GitUpstreamStatus | undefined
+  refreshBranchCompareRef: MutableRefObject<() => Promise<void>>
+}) {
+  const branchCompareStatusHeadRef = useRef<BranchCompareStatusHeadSnapshot | null>(null)
+  const branchCompareRemoteStatusRef = useRef<BranchCompareRemoteStatusSnapshot | null>(null)
+
+  useEffect(() => {
+    if (!activeWorktreeId || !worktreePath || !isBranchVisible || !compareBaseRef || isFolder) {
+      branchCompareStatusHeadRef.current = null
+      return
+    }
+    const current = {
+      baseRef: compareBaseRef,
+      statusHead: activeGitStatusHead,
+      worktreeId: activeWorktreeId
+    }
+    const previous = branchCompareStatusHeadRef.current
+    branchCompareStatusHeadRef.current = current
+    if (shouldRefreshBranchCompareForStatusHead(previous, current)) {
+      void refreshBranchCompareRef.current()
+    }
+  }, [
+    activeGitStatusHead,
+    activeWorktreeId,
+    compareBaseRef,
+    isBranchVisible,
+    isFolder,
+    refreshBranchCompareRef,
+    worktreePath
+  ])
+
+  useEffect(() => {
+    if (!activeWorktreeId || !worktreePath || !isBranchVisible || !compareBaseRef || isFolder) {
+      branchCompareRemoteStatusRef.current = null
+      return
+    }
+    // Why: pushing a branch can move its remote base and ahead count without changing local HEAD, which the HEAD-change effect alone misses.
+    const current = {
+      ahead: remoteStatus?.ahead ?? null,
+      baseRef: compareBaseRef,
+      behind: remoteStatus?.behind ?? null,
+      hasUpstream: remoteStatus?.hasUpstream ?? null,
+      upstreamName: remoteStatus?.upstreamName ?? null,
+      worktreeId: activeWorktreeId
+    }
+    const previous = branchCompareRemoteStatusRef.current
+    branchCompareRemoteStatusRef.current = current
+    if (shouldRefreshBranchCompareForRemoteStatus(previous, current)) {
+      void refreshBranchCompareRef.current()
+    }
+  }, [
+    activeWorktreeId,
+    compareBaseRef,
+    isBranchVisible,
+    isFolder,
+    refreshBranchCompareRef,
+    remoteStatus?.ahead,
+    remoteStatus?.behind,
+    remoteStatus?.hasUpstream,
+    remoteStatus?.upstreamName,
+    worktreePath
+  ])
+}
